Update wasm-tools/wit-bindgen deps (#14199)
* Update wasm-tools/wit-bindgen deps

Keepin-up-to-date.

* Update wit-bindgen dependency

prtest:full

+++ crates/environ/src/compile/module_environ.rs
@@ -820,7 +820,7 @@ impl<'a, 'data> ModuleEnvironment<'a, 'data> {
                 self.validator.code_section_start(&range)?;
                 let cnt = usize::try_from(count).unwrap();
                 self.result.function_body_inputs.reserve_exact(cnt);
-                self.result.debuginfo.wasm_file.code_section_offset = range.start as u64;
+                self.result.debuginfo.wasm_file.code_section_offset = range.start;
             }
 
             Payload::CodeSectionEntry(body) => {
